pyproject.toml
[tool.poetry]
name = "GoB"
version = "0.1.0"
description = ""
authors = ["Masahiro Morinaga <Masahiro.Morinaga@cern.ch>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
einops = "^0.4.1"
attrdict = "^2.0.1"
tqdm = "^4.64.0"
PyYAML = "^6.0"
yapf = "^0.32.0"
matplotlib = "^3.5.2"
GPUtil = "^1.4.0"
mlflow = "^1.27.0"


#jaxlib = {path = "/data/morinaga/wheel/jaxlib-0.3.5-py3-none-any.whl"}

jaxlib = {path = "/data/morinaga/wheel/jaxlib-0.3.22-cp38-cp38-linux_x86_64.whl"}
jax = "0.3.22"
dm-haiku = "^0.0.9"
optax = "^0.1.4"
jmp = "^0.0.2"

ray = "^2.2.0"
cupy-cuda11x = "^11.4.0"
#numba = "^0.56.4"
tensorflow = "^2.11.0"
scikit-learn = "^1.2.0"
numpy = "1.23.3"
tensorflow-datasets = "^4.7.0"
tensorrt = "^8.5.2.2" # To diable many warning message from tensorflow
tensorboard-plugin-profile = "^2.11.1"
tensorboard = "^2.11.2"
uproot-methods = "^0.9.2"
tables = "^3.8.0"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
